refactor: adds components for globeIcon and walletIcon
